refactor: 💡 extract fetch logic to custom fetch
