+++ src/types/customFetch/index.ts
@@ -1,0 +1,10 @@
+export type CustomFetchType<T> = {
+  resource: string;
+  method: 'GET' | 'POST' | 'PUT' | 'DELETE';
+  body?: T;
+};
+
+export type SendResponse = {
+  message?: string;
+  error?: string;
+};
